CampusCookery/requirements.txt
Django>=4.0
Pillow>=8.0.0
